+++ tests/src/rules/no-internal-modules.js
@@ -59,6 +59,34 @@ ruleTester.run('no-internal-modules', rule, {
         allow: [ '**/index{.js,}' ],
       } ],
     }),
+    test({
+      code: 'import a from "./plugin2/thing"',
+      filename: testFilePath('./internal-modules/plugins/plugin.js'),
+      options: [ {
+        forbid: [ '**/api/*' ],
+      } ],
+    }),
+    test({
+      code: 'const a = require("./plugin2/thing")',
+      filename: testFilePath('./internal-modules/plugins/plugin.js'),
+      options: [ {
+        forbid: [ '**/api/*' ],
+      } ],
+    }),
+    test({
+      code: 'import b from "app/a"',
+      filename: testFilePath('./internal-modules/plugins/plugin2/internal.js'),
+      options: [ {
+        forbid: [ 'app/**/**' ],
+      } ],
+    }),
+    test({
+      code: 'import b from "@org/package"',
+      filename: testFilePath('./internal-modules/plugins/plugin2/internal.js'),
+      options: [ {
+        forbid: [ '@org/package/*' ],
+      } ],
+    }),
     // exports
     test({
       code: 'export {a} from "./internal.js"',
@@ -114,6 +142,34 @@ ruleTester.run('no-internal-modules', rule, {
         parser: parser,
       }),
     ]),
+    test({
+      code: 'export * from "./plugin2/thing"',
+      filename: testFilePath('./internal-modules/plugins/plugin.js'),
+      options: [ {
+        forbid: [ '**/api/*' ],
+      } ],
+    }),
+    test({
+      code: 'export * from "app/a"',
+      filename: testFilePath('./internal-modules/plugins/plugin2/internal.js'),
+      options: [ {
+        forbid: [ 'app/**/**' ],
+      } ],
+    }),
+    test({
+      code: 'export { b } from "@org/package"',
+      filename: testFilePath('./internal-modules/plugins/plugin2/internal.js'),
+      options: [ {
+        forbid: [ '@org/package/*' ],
+      } ],
+    }),
+    test({
+      code: 'export * from "./app/index.js";\nexport * from "./app/index"',
+      filename: testFilePath('./internal-modules/plugins/plugin2/internal.js'),
+      options: [ {
+        forbid: [ '**/index.ts' ],
+      } ],
+    }),
   ],
 
   invalid: [
@@ -184,6 +240,70 @@ ruleTester.run('no-internal-modules', rule, {
         },
       ],
     }),
+    test({
+      code: 'import "./app/index.js"',
+      filename: testFilePath('./internal-modules/plugins/plugin2/internal.js'),
+      options: [ {
+        forbid: [ '*/app/*' ],
+      } ],
+      errors: [ {
+        message: 'Reaching to "./app/index.js" is not allowed.',
+        line: 1,
+        column: 8,
+      } ],
+    }),
+    test({
+      code: 'import b from "@org/package"',
+      filename: testFilePath('./internal-modules/plugins/plugin2/internal.js'),
+      options: [ {
+        forbid: [ '@org/**' ],
+      } ],
+      errors: [ {
+        message: 'Reaching to "@org/package" is not allowed.',
+        line: 1,
+        column: 15,
+      } ],
+    }),
+    test({
+      code: 'import b from "app/a/b"',
+      filename: testFilePath('./internal-modules/plugins/plugin2/internal.js'),
+      options: [ {
+        forbid: [ 'app/**/**' ],
+      } ],
+      errors: [ {
+        message: 'Reaching to "app/a/b" is not allowed.',
+        line: 1,
+        column: 15,
+      } ],
+    }),
+    test({
+      code: 'import get from "lodash.get"',
+      filename: testFilePath('./internal-modules/plugins/plugin2/index.js'),
+      options: [ {
+        forbid: [ 'lodash.*' ],
+      } ],
+      errors: [ {
+        message: 'Reaching to "lodash.get" is not allowed.',
+        line: 1,
+        column: 17,
+      } ],
+    }),
+    test({
+      code: 'import "./app/index.js";\nimport "./app/index"',
+      filename: testFilePath('./internal-modules/plugins/plugin2/internal.js'),
+      options: [ {
+        forbid: [ '**/index{.js,}' ],
+      } ],
+      errors: [ {
+        message: 'Reaching to "./app/index.js" is not allowed.',
+        line: 1,
+        column: 8,
+      }, {
+        message: 'Reaching to "./app/index" is not allowed.',
+        line: 2,
+        column: 8,
+      } ],
+    }),
     // exports
     test({
       code: 'export * from "./plugin2/index.js";\nexport * from "./plugin2/app/index"',
@@ -248,6 +368,34 @@ ruleTester.run('no-internal-modules', rule, {
           message: 'Reaching to "debug/node" is not allowed.',
           line: 1,
           column: 19,
+        },
+      ],
+    }),
+    test({
+      code: 'export * from "./plugin2/thing"',
+      filename: testFilePath('./internal-modules/plugins/plugin.js'),
+      options: [ {
+        forbid: [ '**/plugin2/*' ],
+      } ],
+      errors: [
+        {
+          message: 'Reaching to "./plugin2/thing" is not allowed.',
+          line: 1,
+          column: 15,
+        },
+      ],
+    }),
+    test({
+      code: 'export * from "app/a"',
+      filename: testFilePath('./internal-modules/plugins/plugin2/internal.js'),
+      options: [ {
+        forbid: [ '**' ],
+      } ],
+      errors: [
+        {
+          message: 'Reaching to "app/a" is not allowed.',
+          line: 1,
+          column: 15,
         },
       ],
     }),